Мечтая сделать один универсальный код для Supplier и Consumer - упустил очень важные детали реализации, что привело к пагубным последствиям. Когда Supplier разгонялся до максимальных TPS, не взирая на то, что у внешнего ресурса может не быть данных и он в холостую поливает внешний ресур. Так же общая статистика оказалась не возможной, потому что Supplier inputTps должен считать кол-вом запросов к внешнему источнику (до момента получения ответа), а Consumer должен считать inputTps по кол-ву считанных сообщений. Вообщем всё удалил и написал заного Supplier. Consumer ещё не тестировал.

+++ src/main/java/ru/jamsys/scheduler/AbstractThreadBalancerScheduler.java
@@ -11,7 +11,7 @@ import java.util.function.Function;
 
 public abstract class AbstractThreadBalancerScheduler extends AbstractScheduler implements SchedulerTick {
 
-    public AbstractThreadBalancerScheduler(String name, int periodMillis) {
+    public AbstractThreadBalancerScheduler(String name, long periodMillis) {
         super(name, periodMillis);
     }
